App: Set a better secondary color
Instead of setting a grey-ish color as the secondary, which makes it hard to
distinguish components using it from disabled ones, use a non-grey color
instead.

The colors were chosen using https://www.canva.com/colors/color-wheel/, using
the primary color as the basis. I wanted to choose a secondary that does not
clash with the various alert colors.

So for light mode, I chose a blue-ish/purple-ish, darker color, and for dark
mode, a more cyan-ish. They both feel okay with the primary keyboardio-orange.

Fixes #1089.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>