Preferences: Add support for configuring a few OneShot settings
Requires keyboardio/Kaleidoscope#1277 and the `OneShotConfig` plugin on the
firmware side for the new options to appear.

The new options are merged together with the existing `EscapeOneShot` config, as
they all configure various OneShot-related things.

Setting the "Double tap behaviour" is more restricted than what the plugin would
allow: the plugin allows setting the behaviour on a per-key basis, for all
built-in oneshot keys. For the sake of simplicity, Chrysalis only offers a few
preset settings: enabling locking for all oneshot keys, for modifiers only, for
layer keys only, or disabling the behaviour completely.

Addresses the Chrysalis parts of #732.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>