Preferences: Drop the double tap timeout configuration from OneShot for now
Presenting a good UI for the double tap timeout is not trivial, because it's
default value (`-1`) tells the firmware to use the same value as the normal
timeout. Just letting the slider go to -1, and displaying "-1ms" in the label is
wrong.

Until we manage to come up with a better presentation, remove the setting from
the OneShot preferences section. The setting defaults to following the normal
timeout anyway, and that's a very reasonable value for it in the vast majority
of cases.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>